+++ tests/result_srv.py
@@ -4,11 +4,14 @@ success = False
 test_nbr = 0
 number_of_success_expected = 0
 success_counter = 0
+multiple_objectives = False
+done_objectives = []
 
-def set_test(nbr, counter):
-    global success, test_nbr, number_of_success_expected, success_counter
+def set_test(nbr, counter, multiple_obj: bool = False):
+    global success, test_nbr, number_of_success_expected, success_counter, multiple_objectives
     success = False
     test_nbr = nbr
+    multiple_objectives = multiple_obj
     number_of_success_expected = counter
     success_counter = 0
 
@@ -35,6 +38,17 @@ class ResultServer(BaseHTTPRequestHandler):
         if self.path.strip() == f"/r/{test_nbr}/ok":
             success_counter += 1
             if success_counter >= number_of_success_expected:
+                print(f"Success: {success_counter}/{number_of_success_expected}")
                 success = True
+        elif multiple_objectives and self.path.strip().startswith(f"/r/{test_nbr}/"):
+            print(f"Fail: {success_counter}/{number_of_success_expected}")
+        if self.path.strip().startswith(f"/r/{test_nbr}/ok/") and multiple_objectives:
+            server_nbr = self.path.strip().split("/")[-1]
+            if server_nbr not in done_objectives:
+                done_objectives.append(server_nbr)
+                success_counter += 1
+                if success_counter >= number_of_success_expected:
+                    print(f"Success: {success_counter}/{number_of_success_expected}")
+                    success = True
         self.wfile.write(bytes(f"Success: {success_counter}/{number_of_success_expected}", "utf-8"))
         self.end_headers()